(PC-19402) feat(useHomeRecommendedHits): extract getRecommentationEndpoint fucntion into a helper file
